Use preload option instead injecting script everytime
